added comand for bomb decorator & created comands.cpp, comands.h
