+++ SBomberProject/DynamicObject.h
@@ -10,8 +10,8 @@ public:
 
     DynamicObject() : speed(0.0), xDirction(0.0), yDirection(0) { }
 
-    inline void SetSpeed(double sp) { speed = sp; }
-    inline void SetDirection(double dx, double dy) { xDirction = dx; yDirection = dy; }
+    inline virtual void SetSpeed(double sp) { speed = sp; }
+    inline virtual void SetDirection(double dx, double dy) { xDirction = dx; yDirection = dy; }
     
     virtual void Move(uint16_t time) { x += xDirction * speed * time * 0.001; y += yDirection * speed * time * 0.001; };
